refac: Include e2e monitoring stack tests
