+++ Makefile
@@ -38,10 +38,10 @@ test: generate ## run tests
 	@go test -coverprofile=coverage/coverage.out -covermode=count -timeout 25m ./...
 
 e2e-test: generate ## Run e2e tests
-	@go test -timeout 20m -count=1 ./e2e/...
+	@go test -timeout 25m -count=1 ./e2e/...
 
 e2e-test-windows: generate ## Run e2e tests on Windows
-	@go test -timeout 20m -count=1 -skip TestE2E_MonitoringStack ./e2e/...
+	@go test -timeout 25m -count=1 -skip TestE2E_MonitoringStack ./e2e/...
 
 test-no-e2e: generate ## run tests excluding e2e
 	@mkdir -p coverage
