+++ e2e/sedge/monitoring_stack_test.go
@@ -1,8 +1,24 @@
+/*
+Copyright 2022 Nethermind
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+	http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
 package e2e
 
 import (
 	"testing"
 
+	base "github.com/NethermindEth/sedge/e2e"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -13,13 +29,13 @@ func TestE2E_MonitoringStack_NotInitialized(t *testing.T) {
 		runErr error
 	)
 	// Build test case
-	e2eTest := newE2ETestCase(
+	e2eTest := newE2ESedgeTestCase(
 		t,
 		// Arrange
 		nil,
 		// Act
 		func(t *testing.T, binaryPath string, dataDirPath string) {
-			runErr = runCommand(t, binaryPath, "--help")
+			runErr = base.RunCommand(t, binaryPath, "--help")
 		},
 		// Assert
 		func(t *testing.T, dataDirPath string) {
@@ -40,13 +56,13 @@ func TestE2E_MonitoringStack_Init(t *testing.T) {
 		runErr error
 	)
 	// Build test case
-	e2eTest := newE2ETestCase(
+	e2eTest := newE2ESedgeTestCase(
 		t,
 		// Arrange
 		nil,
 		// Act
 		func(t *testing.T, binaryPath string, dataDirPath string) {
-			runErr = runCommand(t, binaryPath, "monitoring", "init")
+			runErr = base.RunCommand(t, binaryPath, "monitoring", "init")
 		},
 		// Assert
 		func(t *testing.T, dataDirPath string) {
@@ -70,11 +86,11 @@ func TestE2E_MonitoringStack_NotReinstalled(t *testing.T) {
 		runErr                  error
 	)
 	// Build test case
-	e2eTest := newE2ETestCase(
+	e2eTest := newE2ESedgeTestCase(
 		t,
 		// Arrange
 		func(t *testing.T, sedgePath string) error {
-			err := runCommand(t, sedgePath, "monitoring", "init")
+			err := base.RunCommand(t, sedgePath, "monitoring", "init")
 			if err != nil {
 				return err
 			}
@@ -91,7 +107,7 @@ func TestE2E_MonitoringStack_NotReinstalled(t *testing.T) {
 		},
 		// Act
 		func(t *testing.T, binaryPath string, dataDirPath string) {
-			runErr = runCommand(t, binaryPath, "monitoring", "init")
+			runErr = base.RunCommand(t, binaryPath, "monitoring", "init")
 		},
 		// Assert
 		func(t *testing.T, dataDirPath string) {
@@ -121,15 +137,15 @@ func TestE2E_MonitoringStack_Clean(t *testing.T) {
 		runErr error
 	)
 	// Build test case
-	e2eTest := newE2ETestCase(
+	e2eTest := newE2ESedgeTestCase(
 		t,
 		// Arrange
 		func(t *testing.T, sedgePath string) error {
-			return runCommand(t, sedgePath, "monitoring", "init")
+			return base.RunCommand(t, sedgePath, "monitoring", "init")
 		},
 		// Act
 		func(t *testing.T, binaryPath string, dataDirPath string) {
-			runErr = runCommand(t, binaryPath, "monitoring", "clean")
+			runErr = base.RunCommand(t, binaryPath, "monitoring", "clean")
 		},
 		// Assert
 		func(t *testing.T, dataDirPath string) {
@@ -152,13 +168,13 @@ func TestE2E_MonitoringStack_CleanNonExistent(t *testing.T) {
 		runErr error
 	)
 	// Build test case
-	e2eTest := newE2ETestCase(
+	e2eTest := newE2ESedgeTestCase(
 		t,
 		// Arrange
 		nil,
 		// Act
 		func(t *testing.T, binaryPath string, dataDirPath string) {
-			runErr = runCommand(t, binaryPath, "monitoring", "clean")
+			runErr = base.RunCommand(t, binaryPath, "monitoring", "clean")
 		},
 		// Assert
 		func(t *testing.T, dataDirPath string) {
